Move example app package
