+++ example/android/app/src/main/java/nabidreams/audiorecorder/example/MainApplication.java
@@ -1,7 +1,8 @@
-package com.audiorecorderexample;
+package nabidreams.audiorecorder.example;
 
 import android.app.Application;
 import android.content.Context;
+
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
